feat: apis 파일 대분류로 통합, models 파일 분리
