+++ src/apis/content/content.ts
@@ -13,3 +13,19 @@ export const getContent = async (contentId: number | undefined): Promise<GetCont
 
   return data.data
 }
+
+interface DeleteContentParams {
+  contentId: number
+}
+
+/**
+ *
+ */
+export const deleteContent = async ({ contentId }: DeleteContentParams) => {
+  const response = await client({
+    method: 'DELETE',
+    url: `/contents/${contentId}`,
+  })
+
+  return response
+}
