--- src/apis/content/deleteContent.ts
@@ -1,17 +1,0 @@
-import { client } from '../client'
-
-interface DeleteContentParams {
-  contentId: number
-}
-
-/**
- *
- */
-export const deleteContent = async ({ contentId }: DeleteContentParams) => {
-  const response = await client({
-    method: 'DELETE',
-    url: `/contents/${contentId}`,
-  })
-
-  return response
-}
