+++ src/models/archiving/ArchivingList.ts
@@ -1,22 +1,3 @@
-export interface HomeArchivingList {
-  content: [
-    {
-      categoryId: number
-      title: string
-      imageUrl: string
-      createdAt: string
-      category: string
-      imgCnt: number
-      linkCnt: number
-      scrapCnt: number
-      markStatus: boolean
-    }
-  ]
-  page: number
-  size: number
-  hasNext: boolean
-}
-
 export interface ArchivingListResponse {
   [category: string]: ArchivingItem[]
 }
