+++ src/models/archiving/ContentByArchiving.ts
@@ -1,0 +1,22 @@
+export interface ContentByArchivingResponse {
+  contents: {
+    content: [
+      {
+        contentId: number
+        contentTitle: string
+        contentType: string
+        link: string
+        imgUrl: string
+        contentCreatedAt: string
+        tag: string
+        tagCount: number
+      }
+    ]
+    page: number
+    size: number
+    hasNext: boolean
+  }
+  archivingTitle: string
+  archivingId: number
+  totalContentsCount: number
+}
