+++ src/screens/contentDetail/ContentDetail.tsx
@@ -6,8 +6,7 @@ import { AxiosError } from 'axios'
 import { Text, SafeAreaView, ScrollView } from 'react-native'
 import { useMutation, useQuery } from 'react-query'
 
-import { deleteContent } from '@/apis/content/deleteContent'
-import { getContent } from '@/apis/content/getContent'
+import { deleteContent, getContent } from '@/apis/content/content'
 import { defaultImages } from '@/assets'
 import DefaultDialog from '@/components/dialogs/defaultDialog/DefaultDialog'
 import TwoButtonDialog from '@/components/dialogs/twoButtonDialog/TwoButtonDialog'
